feat: Trigger filter priority (#1)
I wanted to be sure I understand the exact interplay between the tags, branches and paths filters. So I created a workflow to explore it in a structured way.

The answer is that the workflow runs if: tags || (branches && paths)
If branches or tags are not defined they are true.